third_party/blink/web_tests/platform/generic/external/wpt/css/css-fonts/parsing/font-face-src-tech-expected.txt

This is a testharness.js-based test.
PASS Check that src: url("foo.ttf") is valid
PASS Check that src: url("foo.ttf") tech() is invalid
FAIL Check that src: url("foo.ttf") tech(feature-opentype) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(feature-aat) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(feature-graphite) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-colrv0) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-colrv1) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-svg) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-sbix) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(color-cbdt) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(variations) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(palettes) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(incremental) is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") tech("feature-opentype") is invalid
PASS Check that src: url("foo.ttf") tech("color-colrv0") is invalid
PASS Check that src: url("foo.ttf") tech("variations") is invalid
FAIL Check that src: url("foo.ttf") tech(feature-opentype, color-colrv0, variations, palettes) is valid assert_equals: expected true but got false
FAIL Check that src: url("foo.ttf") tech(incremental, color-svg, feature-graphite, feature-aat) is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") tech(feature-opentype color-colrv0 variations palettes) is invalid
PASS Check that src: url("foo.ttf") tech(auto) is invalid
PASS Check that src: url("foo.ttf") tech(default) is invalid
PASS Check that src: url("foo.ttf") tech(inherit) is invalid
PASS Check that src: url("foo.ttf") tech(initial) is invalid
PASS Check that src: url("foo.ttf") tech(none) is invalid
PASS Check that src: url("foo.ttf") tech(normal) is invalid
PASS Check that src: url("foo.ttf") tech(xyzzy) is invalid
FAIL Check that src: url("foo.ttf") format(opentype) tech(feature-opentype) is valid assert_equals: expected true but got false
PASS Check that src: url("foo.ttf") tech(feature-opentype) format(opentype) is invalid
Harness: the test ran to completion.
